# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Virus Tracker API
  version: 1.0.0
  description: |
    REST interface for the virus evolution tracking platform. Exposes sequence analytics,
    epidemiological metrics, authentication, and administrative capabilities implemented in the
    Rust Axum service. Most endpoints are available under the `/api` prefix.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://virus-tracker.example.com
    description: Production (example)
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Static API key for privileged ingestion endpoints.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        roles:
          type: array
          items:
            type: string
      required: [token, roles]
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        roles:
          type: array
          items:
            type: string
          description: Optional role list, defaults to `viewer`.
      required: [email, password]
    CreateUserResponse:
      type: object
      properties:
        user_id:
          type: string
        roles:
          type: array
          items:
            type: string
      required: [user_id, roles]
    UserSummary:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time
        roles:
          type: array
          items:
            type: string
      required: [id, email, created_at, roles]
paths:
...
  /api/auth/login:
    post:
      summary: Authenticate and retrieve JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
  /api/admin/users:
    get:
      summary: List users and roles
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Collection of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummary'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a new user with optional role assignments
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden